Refactor the code removing unnecessary methods.
